+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/faultinjection/FaultInjectionServerErrorRuleOnDirectTests.java
@@ -1288,6 +1288,70 @@ public class FaultInjectionServerErrorRuleOnDirectTests extends FaultInjectionTe
         }
     }
 
+    @Test(groups = {"long"}, timeOut = TIMEOUT)
+    public void faultInjectionInjectTcpResponseDelay() throws JsonProcessingException {
+        CosmosAsyncClient newClient = null; // creating new client to force creating new connections
+        // define another rule which can simulate timeout
+        String timeoutRuleId = "serverErrorRule-transitTimeout-" + UUID.randomUUID();
+        FaultInjectionRule timeoutRule =
+            new FaultInjectionRuleBuilder(timeoutRuleId)
+                .condition(
+                    new FaultInjectionConditionBuilder()
+                        .operationType(FaultInjectionOperationType.CREATE_ITEM)
+                        .build()
+                )
+                .result(
+                    FaultInjectionResultBuilders
+                        .getResultBuilder(FaultInjectionServerErrorType.RESPONSE_DELAY)
+                        .times(1)
+                        .delay(Duration.ofSeconds(4))
+                        .build()
+                )
+                .duration(Duration.ofMinutes(5))
+                .build();
+        try {
+            DirectConnectionConfig directConnectionConfig = DirectConnectionConfig.getDefaultConfig();
+
+            // set networkRequestTimeout to 1.1s but CosmosClient internally sets networkRequestTimeout back to 5s, so the injected delay (4s) will not cause failures
+            directConnectionConfig.setNetworkRequestTimeout(Duration.ofMillis(1100));
+
+            newClient = new CosmosClientBuilder()
+                .endpoint(TestConfigurations.HOST)
+                .key(TestConfigurations.MASTER_KEY)
+                .contentResponseOnWriteEnabled(true)
+                .consistencyLevel(BridgeInternal.getContextClient(this.clientWithoutPreferredRegions).getConsistencyLevel())
+                .directMode(directConnectionConfig)
+                .buildAsyncClient();
+
+            CosmosAsyncContainer container =
+                newClient
+                    .getDatabase(cosmosAsyncContainer.getDatabase().getId())
+                    .getContainer(cosmosAsyncContainer.getId());
+
+            CosmosFaultInjectionHelper.configureFaultInjectionRules(container, Arrays.asList(timeoutRule)).block();
+
+            // create a new item to be used by read operations
+            TestObject createdItem = TestObject.create();
+            CosmosItemResponse<TestObject> itemResponse = container.createItem(createdItem).block();
+
+            assertThat(timeoutRule.getHitCount()).isEqualTo(1);
+            this.validateHitCount(timeoutRule, 1, OperationType.Create, ResourceType.Document);
+
+            this.validateFaultInjectionRuleApplied(
+                itemResponse.getDiagnostics(),
+                OperationType.Create,
+                HttpConstants.StatusCodes.CREATED,
+                HttpConstants.SubStatusCodes.UNKNOWN,
+                timeoutRuleId,
+                false
+            );
+
+        } finally {
+            timeoutRule.disable();
+            safeClose(newClient);
+        }
+    }
+
     private void validateFaultInjectionRuleApplied(
         CosmosDiagnostics cosmosDiagnostics,
         OperationType operationType,